% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datalabR-package.R
\docType{package}
\name{datalabR-package}
\alias{datalabR}
\alias{datalabR-package}
\title{datalabR: Useful tools from the Data Lab ICMAN}
\description{
Provides a suite of useful tools from the Data Lab ICMAN.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datalab-icman/datalabR}
  \item \url{https://datalab-icman.github.io/datalabR}
}

}
\author{
\strong{Maintainer}: Pablo Almaraz \email{datalab.icman@csic.es}

}
\keyword{internal}
