% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_symbol_legend.R
\name{show_symbol_legend}
\alias{show_symbol_legend}
\title{Displays a legend of status and progress symbols}
\usage{
show_symbol_legend()
}
\value{
Prints the messages to the console.
}
\description{
Prints a legend of common symbols used in R package console
messages, using their Unicode escape codes.
}
\examples{
\dontrun{
show_symbol_legend()
}

}
