% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wflow_datalabr.R
\name{wflow_datalabr}
\alias{wflow_datalabr}
\title{A simple custom wrapper to create, configure and tune a workflowr project}
\usage{
wflow_datalabr(
  project_name,
  existing = FALSE,
  username,
  creator,
  domain = NULL,
  repo = c("GitHub", "GitLab"),
  attribution,
  lic = c("CCBY-NC-SA4.0", "gnu-gpl-v3")
)
}
\arguments{
\item{project_name}{Character. The name of the project and repository.}

\item{existing}{Character. Checks if the package already exists.}

\item{username}{Character. Your username on the GitLab/GitHub instance.}

\item{creator}{Character. The name of the author or entity creating the project.}

\item{domain}{Character. The domain of the GitLab/GitHub instance (e.g., "git.csic.es").}

\item{repo}{Character. The type of the repository to commit. Must be one of
c("GitHub", "GitLab")}

\item{attribution}{Character. Attribution text for the license file.}

\item{lic}{Character. The license to apply. Must be one of "CCBY-NC-SA4.0" (default)
or "gnu-gpl-v3".}
}
\value{
Invisibly returns the project path upon successful creation.
}
\description{
This function automates the creation of a workflowr project.
It initializes the project, configures GitLab/GitHub, creates and overwrites
specific R Markdown files with predefined templates, replaces placeholders,
and publishes the initial site.
}
\examples{
\dontrun{
wflow_datalabr(
  project_name = "my_awesome_project",
  username = "Data Lab ICMAN",
  creator = "Data Lab ICMAN",
  domain = "git.csic.es",
  repo = "GitLab",
  attribution = "Data Lab ICMAN",
  lic = "CCBY-NC-SA4.0"
)
}
}
